It seems like the top bits of a blockSize are sometimes used for storing other data, so I'm masking that for the moment (this was found in MAST of the stack templates).
There is a partContentType=128 that appears to have extra data stored in front of the text.  No clue what that data is, but I think I am able to find the offset to that data just fine, so we have a working hack for now that fills up a lot more text boxes/etc with good data and makes our sample stacks look a lot nicer.